Image Content Filtering/requirements.txt
tensorflow
numpy
pillow
opencv-python
requests
